+++ CHANGELOG.md
@@ -5,10 +5,13 @@
 ### Breaking Changes
 
 ### Added
+- Added list keyboard navigation and shortcuts (selection, edit, complete, add, move, delete, search focus). ([#20](https://github.com/kcosr/assistant/pull/20))
 
 ### Changed
+- Added global setting for list single-click selection and refined click behavior. ([#20](https://github.com/kcosr/assistant/pull/20))
 
 ### Fixed
+- Fixed notes/lists context to use the active item instance id. ([#20](https://github.com/kcosr/assistant/pull/20))
 
 ### Removed